Added SweetAlert To AllReview Page

+++ src/pages/Dashboard/ModeratorDashboard/AllReviews/AllReviews.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import useAxios from '../../../../hooks/useAxios';
 import { useQuery } from '@tanstack/react-query';
+import Swal from 'sweetalert2';
 
 const AllReviews = () => {
 
@@ -16,11 +17,29 @@ const AllReviews = () => {
     refetch();
 
     const deleteReview = review => {
-        axios.delete(`/review/${review._id}`)
-            .then(res => {
-                refetch();
-                console.log('delete scholarship success', res.data);
-            })
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`/review/${review._id}`)
+                    .then(res => {
+                        if (res.data.deletedCount) {
+                            refetch();
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your parcel request has been deleted.",
+                                icon: "success"
+                            });
+                        }
+                    })
+            }
+        });
     }
 
     return (
